+++ src/createTelemetryHandler.ts
@@ -1,0 +1,22 @@
+import { ApplicationInsightsV2TelemetryHandler } from './ApplicationInsightsV2TelemetryHandler';
+import { ApplicationInsightsV3TelemetryHandler } from './ApplicationInsightsV3TelemetryHandler';
+import type { CreateTelemetryHandlerOptions, TelemetryHandler } from './types';
+
+export const createTelemetryHandler = (options: CreateTelemetryHandlerOptions): TelemetryHandler => {
+  switch (options.version) {
+    case 2: {
+      return new ApplicationInsightsV2TelemetryHandler({
+        client: options.client,
+        traceFilter: options.traceFilter,
+        exceptionFilter: options.exceptionFilter,
+      });
+    }
+    case 3: {
+      return new ApplicationInsightsV3TelemetryHandler({
+        client: options.client,
+        traceFilter: options.traceFilter,
+        exceptionFilter: options.exceptionFilter,
+      });
+    }
+  }
+};