+++ src/index.ts
@@ -1,10 +1,10 @@
-import { ApplicationInsightsTransport } from './ApplicationInsightsTransport';
-import { createApplicationInsightsTransport, createTelemetryHandler, createWinstonLogger } from './createWinstonLogger';
+import { createApplicationInsightsTransport } from './createApplicationInsightsTransport';
+import { createTelemetryHandler } from './createTelemetryHandler';
+import { createWinstonLogger } from './createWinstonLogger';
 import { TelemetrySeverity } from './enums';
 import { isRunningInAzure } from './isRunningInAzure';
 import { isRunningLocally } from './isRunningLocally';
-import type { ITelemetryFilterV2, ITelemetryFilterV3 } from './types';
-import type { AzureApplicationInsightsLoggerOptions, ExtractedProperties, RequiredOptions, SeverityMapping, TelemetryData, TelemetryHandler, WinstonInfo } from './types';
+import type { AzureApplicationInsightsLoggerOptions, CreateApplicationInsightsTransportOptions, CreateTelemetryHandlerOptions, CreateWinstonLoggerOptions, ITelemetryFilterV2, ITelemetryFilterV3, TelemetryHandler, WinstonInfo } from './types';
 
 export { createApplicationInsightsTransport, createTelemetryHandler, createWinstonLogger, TelemetrySeverity, isRunningInAzure, isRunningLocally };
-export type { ITelemetryFilterV2, ITelemetryFilterV3, TelemetryHandler, WinstonInfo, AzureApplicationInsightsLoggerOptions };
+export type { ITelemetryFilterV2, ITelemetryFilterV3, TelemetryHandler, WinstonInfo, AzureApplicationInsightsLoggerOptions, CreateApplicationInsightsTransportOptions, CreateTelemetryHandlerOptions as TelemetryHandlerFactoryOptions, CreateWinstonLoggerOptions };